requirements.txt
openai
langchain
faiss-cpu
pypdf
unstructured
python-dotenv
tqdm
sentence-transformers
streamlit
fastapi
uvicorn